Switch ethsore to new crypto
